Trying to refactor repo to be a ESP-IDF component.
