+++ HD44780/CMakeLists.txt
@@ -1,0 +1,5 @@
+idf_component_register(
+    SRCS "HD44780.c"
+    INCLUDE_DIRS "include"
+    REQUIRES "driver" "rom" "freertos"
+    )
