+++ HD44780/examples/HD44780_test_scroll/CMakeLists.txt
@@ -5,4 +5,4 @@
 cmake_minimum_required(VERSION 3.16)
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-project(Hitachi_LCD_Test)
+project(HD44780)
