--- components/HD44780/CMakeLists.txt
@@ -1,2 +1,0 @@
-idf_component_register(SRCS "HD44780.c"
-                    INCLUDE_DIRS "include")
